pyproject.toml
[project]
name = "handypi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mediapipe>=0.10.21",
    "numpy>=1.26.4",
    "opencv-python>=4.11.0.86",
    "pika>=1.3.2",
    "python-dotenv>=1.0.0",
    "rich>=14.2.0",
]
